+++ model/user.go
@@ -2,7 +2,7 @@ package model
 
 //easyjson:json
 type User struct {
-	Id       int    `json:"userId"`
+	Id       uint64 `json:"userId"`
 	Login    string `json:"login"`
 	Password string `json:"password"`
 	Name     string `json:"name"`
